requirements.txt
Django==5.0
PyPDF2==3.0.1
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.7
pdf2image==1.16.3
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
whitenoise==6.6.0
